5.0-python/2.2-login-api/src/requirements.txt
Flask==1.1.2
flask-sqlalchemy-2.4.4
flask-marshmallow-0.14.0
marshmallow-sqlalchemy-0.24.1
pymysql-1.0.2
